requirements.txt
# ============================================================================
# HOMELAB AUTONOMOUS AGENT SYSTEM - PYTHON DEPENDENCIES
# ============================================================================

# ----------------------------------------------------------------------------
# LLM & AI Frameworks
# ----------------------------------------------------------------------------
anthropic>=0.41.0,<1.0
langchain==0.3.13
langchain-anthropic==0.3.3
langchain-community==0.3.13
langgraph==0.2.53
langgraph-checkpoint-postgres==2.0.11

# Agent Memory
mem0ai==0.1.33

# MCP (Model Context Protocol)
mcp==1.1.2

# ----------------------------------------------------------------------------
# Database & Storage
# ----------------------------------------------------------------------------
psycopg2-binary==2.9.10
psycopg-binary>=3.2.11  # PostgreSQL adapter with binary support (requires libpq-dev)
pgvector==0.3.6
redis==5.2.1
sqlalchemy==2.0.36

# ----------------------------------------------------------------------------
# HTTP & API Clients
# ----------------------------------------------------------------------------
aiohttp==3.11.11
requests==2.32.3
httpx==0.28.1

# ----------------------------------------------------------------------------
# Telegram Bot
# ----------------------------------------------------------------------------
python-telegram-bot==21.9

# ----------------------------------------------------------------------------
# Prometheus Metrics
# ----------------------------------------------------------------------------
prometheus-client==0.21.1

# ----------------------------------------------------------------------------
# Configuration & Environment
# ----------------------------------------------------------------------------
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# ----------------------------------------------------------------------------
# Proxmox API
# ----------------------------------------------------------------------------
proxmoxer==2.1.0

# ----------------------------------------------------------------------------
# Homelab Service APIs
# ----------------------------------------------------------------------------
# Tailscale
# tailscale-api==0.1.0  # Unofficial - using requests directly instead

# Cloudflare
cloudflare==2.20.0

# Unifi Controller
pyunifi==2.21

# Pi-hole
# pihole-api==1.0.0  # Unofficial - using requests directly instead

# Traefik (via Docker API or file watching)
docker==7.1.0

# Netbox
pynetbox==7.4.1

# MinIO (S3-compatible)
minio==7.2.11

# MQTT
paho-mqtt==2.1.0

# InfluxDB
influxdb-client==1.47.0

# ----------------------------------------------------------------------------
# Kubernetes Client (if K8s enabled)
# ----------------------------------------------------------------------------
kubernetes==31.0.0

# ----------------------------------------------------------------------------
# GitOps & Version Control
# ----------------------------------------------------------------------------
GitPython==3.1.43
PyGithub==2.5.0
python-gitlab==5.2.0

# ----------------------------------------------------------------------------
# Backup & Cloud Storage
# ----------------------------------------------------------------------------
boto3==1.35.94  # AWS S3 / compatible
b2sdk==2.7.0    # Backblaze B2

# ----------------------------------------------------------------------------
# Monitoring & Logging
# ----------------------------------------------------------------------------
structlog==24.4.0
python-json-logger==3.2.1

# ----------------------------------------------------------------------------
# Task Scheduling & Job Queue
# ----------------------------------------------------------------------------
celery==5.4.0
apscheduler==3.10.4

# ----------------------------------------------------------------------------
# Utilities
# ----------------------------------------------------------------------------
pyyaml==6.0.2
toml==0.10.2
click==8.1.8
rich==14.0.0  # Beautiful CLI output
tabulate==0.9.0

# ----------------------------------------------------------------------------
# Testing & Development
# ----------------------------------------------------------------------------
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
pytest-mock==3.14.0
black==24.10.0
flake8==7.1.1
mypy==1.14.1
pylint==3.3.3

# ----------------------------------------------------------------------------
# Security
# ----------------------------------------------------------------------------
cryptography==44.0.0
pyjwt==2.10.1

# ----------------------------------------------------------------------------
# Performance & Optimization
# ----------------------------------------------------------------------------
uvloop==0.21.0  # Faster event loop
orjson==3.10.13  # Faster JSON

# ----------------------------------------------------------------------------
# Networking
# ----------------------------------------------------------------------------
paramiko==3.5.0  # SSH client
fabric==3.2.2    # SSH automation

# ----------------------------------------------------------------------------
# Date & Time
# ----------------------------------------------------------------------------
python-dateutil==2.9.0
pytz==2024.2

# ============================================================================
# Optional Dependencies (install as needed)
# ============================================================================

# Local LLM (Ollama)
# ollama==0.4.4

# Home Assistant
# homeassistant-api==4.1.3

# n8n Python Client
# n8n-client==0.2.0

# Terraform Python wrapper
# python-terraform==0.10.1

# Ansible Python wrapper
# ansible==10.7.0
# ansible-core==2.17.7

# Advanced observability
# opentelemetry-api==1.29.0
# opentelemetry-sdk==1.29.0
# jaeger-client==4.8.0

# Cost tracking
# openai-cost-calculator==1.0.0
